requirements.txt
app==0.0.1
python-dotenv==1.0.1
Requests==2.32.3
SQLAlchemy==1.4.39
pandas==2.0.3
matplotlib==3.7.2
numpy==1.24.3
seaborn==0.12.2
pymysql==1.1.1
mysql-connector-python==9.1.0